+++ src/app/Message/ApplicationMessage.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace PostsDMS\Message;
 
 use Chassis\Framework\Adapters\Message\ApplicationMessageInterface;
@@ -16,8 +18,19 @@ class ApplicationMessage implements ApplicationMessageInterface
      */
     public function __construct(array $body, array $headers = [])
     {
-        $this->payload["items"] = $body;
+        $this->payload["items"] = $this->isArrayAssociative($body) ? [$body] : $body;
         $this->headers = $headers;
+    }
+
+    /**
+     * @param array $array
+     *
+     * @return bool
+     */
+    public function isArrayAssociative(array $array): bool
+    {
+        $keys = array_keys($array);
+        return $keys !== array_keys($keys);
     }
 
     /**