+++ src/app/Providers/MessageRoutingServiceProvider.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace PostsDMS\Providers;
 
 use Chassis\Framework\Providers\RoutingServiceProvider;
-use PostsDMS\OutboundAdapters\GetAuthorInfoData;
+use PostsDMS\OutboundAdapters\GetAuthorData;
 use PostsDMS\OutboundAdapters\PostCreatedEvent;
 use PostsDMS\OutboundAdapters\PostDeletedEvent;
 use PostsDMS\OutboundAdapters\PostUpdatedEvent;
+use PostsDMS\Services\AuthorEventsService;
 use PostsDMS\Services\PostsService;
 
 class MessageRoutingServiceProvider extends RoutingServiceProvider
@@ -17,17 +18,27 @@ class MessageRoutingServiceProvider extends RoutingServiceProvider
      * @var array|string[]
      */
     protected array $inboundRoutes = [
+        // commands
         'getPost' => [PostsService::class, 'get'],
         'createPost' => [PostsService::class, 'create'],
         'updatePost' => [PostsService::class, 'update'],
         'deletePost' => [PostsService::class, 'delete'],
+
+        // events
+        'authorCreated' => AuthorEventsService::class,
+        'authorUpdated' => AuthorEventsService::class,
+        'authorDeleted' => AuthorEventsService::class,
+
     ];
 
     /**
      * @var array|string[]
      */
     protected array $outboundRoutes = [
-        'getAuthorInfo' => GetAuthorInfoData::class,
+        // commands
+        'getAuthor' => GetAuthorData::class,
+
+        // events
         'postCreated' => PostCreatedEvent::class,
         'postUpdated' => PostUpdatedEvent::class,
         'postDeleted' => PostDeletedEvent::class,