+++ src/app/Validator/GetPostRules.php
@@ -11,12 +11,9 @@ class GetPostRules extends AbstractMessageValidator
     // https://github.com/rakit/validation/tree/v1.4.0#available-rules
     protected array $rules = [
         'payload.pathParams.postId' => self::UUID_REGEX_RULE,
-        'headers.jobId' => self::UUID_REGEX_RULE,
     ];
     protected array $messages = [
-        'payload.pathParams.postId:required' => 'postId path property is required',
-        'payload.pathParams.postId' => 'postId path property format is not valid',
-        'headers.jobId:required' => 'jobId header property is required',
-        'headers.jobId' => 'jobId header property format is not valid',
+        'payload.pathParams.postId:required' => 'post id is required',
+        'payload.pathParams.postId' => 'post id format is not valid',
     ];
 }