+++ CHANGELOG.md
@@ -6,7 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.0] - 2022-04-08
+## [0.3.0] - 2022-04-09
+
+### Added
+
+- Add application event message handler
+- Create outbound adapters
+  - redis repository
+
+## [0.2.0] - 2022-04-08
 
 ### Added